Create test_cli test suite
